+++ app/src/main/java/net/skyscanner/backpack/demo/stories/DialogStory.kt
@@ -94,6 +94,7 @@ class DialogStory : Story() {
           R.color.bpkPanjin
         )
         setCancelable(false)
+        setCanceledOnTouchOutside(false)
         setOnCancelListener {
           println("canceled")
         }